[DUOS-1668] split out signing official console tabs into separate pages.
