+++ src/components/DuosHeader.js
@@ -527,7 +527,13 @@ class DuosHeader extends Component {
       },
       isSigningOfficial && {
         label: 'SO Console',
-        link: '/signing_official_console'
+        link: '/signing_official_console/researchers',
+        children: [
+          { label: 'Researchers', link: '/signing_official_console/researchers' },
+          { label: 'DAR Requests', link: '/signing_official_console/dar_requests' },
+          { label: 'Data Submitters', link: '/signing_official_console/data_submitters' },
+          { label: 'DAR Collections', link: '/signing_official_console/dar_collections' }
+        ]
       },
       isResearcher && {
         label: 'Researcher Console',
